Fix bug in indentation of "private def" methods
There was a bug in the plugin's auto-indentation where a method declared
as private on the same line as its definition would fail to indent
correctly.

"private def" was added as a keyword in the regex to keep the regex
simple and explicit. Reading the code shows you that "private def" is
something for which the editor should increase the indentation level
automatically.

Fixes rubyide/vscode-ruby#203