Bugfix: Call the async-callback to clear the parseQueue of items.
Not calling the callback caused a bug that meant that after a few tasks
were enqueued on the parseQueue (i.e. a few change-/create-events), all
incoming tasks were queued and never executed. That lead to newly added
symbols not being added to the symbols list and the locations of all
symbols after that being off.