Fix automatic Rake task detection.
Previously, the globPattern used in rake.ts was failing to find Rakefiles
if they were named 'Rakefile', without the .rb extension. This fixes the
globPattern so now it matches in that case.

This has the downside of also matching any filename that starts with
Rakefile, e.g. 'Rakefile2' or 'Rakefileadsadadsad', but I'm not sure of
a great way to fix this without that downside.