Handle situations where a workspace root folder is undefined. Resolves #533
This can occur when new files are created and assigned as Ruby files. As the file does not exist on the file system, the workspace root folder cannot be determined. First, fall back to the whole project "root" if that's available. Otherwise, allow the language server to just do nothing.