Add Stdin-filename arg to Reek linter
Passes an `--stdin-filename` argument to Reek when linting files.

This gives Reek context about the piped contents and supports using the [`directories` configuration](https://github.com/troessner/reek#working-with-rails).

Caveat:
Because this change sends the absolute path to the file, the configuration in directories must be prefixed with `**` for it to work. e.g.
```yml
directories:
  '**/app/controllers':
    IrresponsibleModule:
      enabled: false
```