Get modern rubyfmt working
Per #627, the previous behavior of the rubyfmt formatter class included
an embedded version of the rubyfmt script. This isn't really a workable
solution anymore as the program has evolved beyond a simple ruby script
to include native modules.  This instead changes the formatter class to
rely on a `rubyfmt` executable on the user's `$PATH`.  Because rubyfmt
is not currently installable via bundler and is in a pre-release state,
users are expected to have followed the necessary installation
instructions from the [project
README](https://github.com/penelopezone/rubyfmt/blob/master/README.md#how-do-i-use-it).
The current instructions recommend adding an alias, but vscode will not
respect that (or at least vscode-ruby doesn't). Instead I have a
[commit](https://github.com/penelopezone/rubyfmt/pull/201/commits/19d5946a43b77de4be1083b3097b632dd79ab1bb)
submitted in [a PR](https://github.com/penelopezone/rubyfmt/pull/201)
which will make it easier to install an actual executable shim script.