feat: add skipFiles and showDebuggerCommands
At Stripe, we use sorbet and other internal utilities that 'pollute' stack frames - they create
a large number of frames that are not interesting to our developers when debugging. This commit
adds two features that help avoid those frames.

skipFiles was inspired by the corresponding js debugger feature. It's useful for avoiding
uninteresting frames that are between interesting frames (sorbet falls in to this category).
Skipped files do not appear in stack traces, an the debugger automatically steps through them.

finishFiles is useful for avoiding frames that define the top of a never-interesting call stack.
The debugger automatically steps out of them instead of through them.

showDebuggerCommands is like showDebuggerOutput, but in the other direction. I found it useful
when debugging both features.